Allow customers to provide custom validation sets for early stopping (#21)
Summary:
Pull Request resolved: https://github.com/facebookincubator/MCGrad/pull/21

Initial context: D87953099

We are following the redesign: the customer can provide the df_val to the fit interface, where the validation set is processed once.

The prev approach heavily relied on repurposing the train_data via is_valid indexing - but sometimes leaned on temp variables of various shapes and forms. I got a bit too confused. So I made some changes in how prediction variables are named and make it explicit which predictions were associated with the training set and which are associated with the validation set.

Reviewed By: TaXxER, flinder

Differential Revision: D87990180

fbshipit-source-id: 385de7a49ba976037e850f8a303b099cd6a29d0a