Improve unit test coverage based on coverage report - plotting.py (#33)
Summary:
This diff adds a bunch of unit tests based on the test coverage report for plotting.py, which had only 21% test coverage.

To be fair, most tests are very shallow, and are essentially a smoke test that checks only if the methods runs without errors.

Adding these tests did bring to initially result in a failed test un, and that uncovered a minor bug:
`plot_calibration_curve` calculates a multicalibration_error, and that method in metrics.py did not support not passing any segments columns. It is a bit silly to not support rendering a calibration curve globally, so I patched that metric.

Update: the Github CI failed because it runs a newer version of sklearn, where the `probas_pred` argument of the PR curve method got renamed to `y_score`. In both versions it is identical as positional argument, and it is unlikely that sklearn will make a change as breaking as having labels as first and scores as second argument. So we switch to using positional arguments to be version-agnostic for sklearn.

Pull Request resolved: https://github.com/facebookincubator/MCGrad/pull/33

Reviewed By: flinder

Differential Revision: D88631943

Pulled By: TaXxER

fbshipit-source-id: 0f0cb3835e9b40a0ae699c36b5ef62d29616e354